requirements.txt
google-generativeai
pandas
scikit-learn
streamlit